Try resolve sbt plugin from valid Maven path
sbt plugins were published to an invalid path with regard to Maven's
specifictation. For instance, 'org/example_2.12_1.0/1.0.0/example-1.0.0.pom'
is invalid because the file name 'example-1.0.0.pom' does not start with
'example_2.12_1.0', the name of the directory that should correspond
to the artifact ID.

As of sbt 1.9, we also publish sbt plugins to the valid path:
example_2.12_1.0/1.0.0/example_2.12_1.0-1.0.0.pom

For backward compatibility, Coursier should try to resolve sbt plugins
from the new Maven path, and fallback to the deprecated Maven path.